updated pagination loading

+++ src/main/java/com/example/chats/ChatMessageRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ChatMessageRepository extends JpaRepository<MessageEntity, Long> {
+public interface ChatMessageRepository extends JpaRepository<MessageEntity, Integer> {
     Page<MessageEntity> findByLocationId(Long locationId, Pageable pageable);
 }
